fixed file delete issue when a user deletes a file. Fix: allow access for redirecting unauthenticated user

+++ src/pages/files/[slug].tsx
@@ -9,6 +9,18 @@ import { convertTimestamp, formatFileSize, isImageFile, isPDFFile } from '@/cons
 import Image from 'next/legacy/image';
 import { BiCopy } from 'react-icons/bi';
 import SocialShare from '@/components/ui/cards/ShareSocial';
+import {
+    FacebookShareButton,
+    FacebookIcon,
+    TwitterShareButton,
+    TwitterIcon,
+    LinkedinShareButton,
+    LinkedinIcon,
+    WhatsappShareButton,
+    WhatsappIcon,
+} from 'next-share';
+import Modal from '@/components/ui/modal/Modal';
+import { div } from 'framer-motion/client';
 
 const Slug = () => {
     const db = getFirestore(app);
@@ -33,6 +45,7 @@ const Slug = () => {
     // Filter files to find the selected one
     const filteredFiles = files.filter((file: { data: { email: string | undefined; }; }) => file.data.email === user?.primaryEmailAddress?.emailAddress);
     const selectedFile = filteredFiles.find((file: { data: { id: string | string[] | undefined; }; }) => file.data.id === slug);
+    const url = selectedFile?.data?.shortURL
 
     // Determine the file type
     useEffect(() => {
@@ -54,17 +67,26 @@ const Slug = () => {
     };
 
     // Delete from Firebase
-    const deleteItem = async (collectionName: string, docId: string) => {
+    const deleteItem = async (collectionName: string, docId: string | string[] | undefined) => {
+        if (!docId || typeof docId !== 'string') {
+            console.error('Invalid document ID:', docId);
+            return;
+        }
+
         try {
             const docRef = doc(db, collectionName, docId);
             await deleteDoc(docRef);
             alert('Document deleted successfully');
-            router.push('/files/');
-            deleteBookmark(docId);
+            // deleteBookmark(docId); // Assuming this removes it from local state/context
+            router.replace('/files').then(() => {
+                // Once the navigation is complete, reload the new page
+                router.reload();
+            });
         } catch (error) {
             console.error('Error deleting document:', error);
         }
     };
+
 
     if (!selectedFile) {
         return (
@@ -84,8 +106,8 @@ const Slug = () => {
                     <span className='text-sm'>{selectedFile.data.fileName}</span>
                 </h1>
             </div>
-            <div className='grid grid-cols-1 sm:grid-cols-2 w-full h-[600px] items-center justify-center p-4 gap-4'>
-                <div className="w-full border h-full relative">
+            <div className='grid grid-cols-1 sm:grid-cols-2 w-full h-[600px] items-center justify-center p-4 gap-4 relative'>
+                <div className="w-full border h-full">
                     {fileType === 'image' && (
                         <Image src={selectedFile.data.fileURL} alt="Uploaded File" layout='fill' objectFit='cover' />
                     )}
@@ -106,19 +128,38 @@ const Slug = () => {
 
                     {/* {!fileType && <p>Unsupported file type or file not found.</p>} */}
                 </div>
-                <div className="w-full border p-1 border-slate-200 h-full">
+                <div className="w-full border p-1 border-slate-200 h-full relative" >
                     <p className='w-full py-2 flex items-center justify-start'><p className='text-black font-bold w-full'>Email Address:</p> <span className='text-left inline-block w-full'>{selectedFile?.data?.email}</span></p>
                     <p className='w-full py-2 flex items-center justify-start'><span className='text-black font-bold w-full'>File Name:</span> <span className='text-left inline-block w-full'>{selectedFile?.data?.fileName}</span></p>
                     <p className='w-full py-2 flex items-center justify-start'><span className='text-black font-bold w-full'>File Size:</span> <span className='text-left inline-block w-full'>{formatFileSize(selectedFile?.data?.fileSize)}</span></p>
                     <p className='w-full py-2 flex items-center justify-start'><span className='text-black font-bold w-full'>File Type:</span> <span className='text-left inline-block w-full'>{selectedFile?.data?.fileType.slice(-15)}</span></p>
                     <p className='w-full py-2 flex items-center justify-start'><span className='text-black font-bold w-full'>Last Modified:</span> <span className='text-left inline-block w-full'>{convertTimestamp(selectedFile?.data?.createdAt)}</span></p>
                     <p className='w-full py-2 flex items-center justify-start'><span className='text-black font-bold w-full'>File URL</span> <span className='text-left inline-block w-full'>{selectedFile?.data?.shortURL} <BiCopy onClick={() => handleCopy(selectedFile?.data?.shortURL)} /> </span></p>
-                    <p className='w-full flex items-center justify-between gap-2'><button className='w-full text-white bg-blue-600 p-2' onClick={() => setShare(true)}>Share</button> <button onClick={() => deleteItem('upload', selectedFile.data.id)} className='text-red-600 bg-red-200 w-full p-2'>Delete</button></p>
+                    <p className='w-full flex items-center justify-between gap-2 outline-none'><button className='w-full text-white bg-blue-600 p-2 outline-none' onClick={() => setShare(true)}>Share</button> <button onClick={() => deleteItem('upload', selectedFile?.id)} className='text-red-600 bg-red-200 w-full p-2'>Delete</button></p>
+                    {share && (
+                        <div className='w-full absolute p-4 rounded-xl h-[150px] flex items-center justify-center flex-col'>
+                            <div className="w-full h-full bg-blue-100 z-10 rounded-t-xl flex items-center  justify-center gap-3">
+                                <FacebookShareButton url={url}>
+                                    <FacebookIcon size={32} round />
+                                </FacebookShareButton>
+                                <TwitterShareButton url={url}>
+                                    <TwitterIcon size={32} round />
+                                </TwitterShareButton>
+                                <LinkedinShareButton url={url}>
+                                    <LinkedinIcon size={32} round />
+                                </LinkedinShareButton>
+                                <WhatsappShareButton url={url}>
+                                    <WhatsappIcon size={32} round />
+                                </WhatsappShareButton>
+                            </div>
+                            <button onClick={() => setShare(false)} className="ml-auto text-red-500 w-full bg-red-100 p-2 rounded-b-lg outline-none">Close</button>
+                        </div>
+                        // <div className='w-full h-[400px] bg-red-600'>
+
+                        // </div>
+                    )}
                 </div>
             </div>
-            {share && (
-                <SocialShare url={selectedFile?.data?.shortURL} />
-            )}
         </Layout>
     );
 };
